requirements.txt
dify-plugin~=0.7.2
httpx>=0.27
numpy>=1.24
zstandard>=0.22
faiss-cpu>=1.7
sentence-transformers>=2.2
